Removed `defaultNodes`
